requirements.txt
Django==1.9.5
argparse==1.2.1
django-rest==0.0.1
djangorestframework==3.3.3
psycopg2==2.6.1
wsgiref==0.1.2
